Add files via upload
Music genre classification is consisted of two main components, audio features and classifier. This paper implemented a classical long short term memory (LSTM) model on music genre classification on GTZAN dataset and proposed a paral- lel CNN and RNN model to improve the performance by using various kinds of audio features. After trials on segmentation of the audio files, the proposed Parallel CNN and RNN model reached an accuracy of 72.5%. We found that the architecture of combining CNN and GRU takes into account of both time series feature and melspectrogram image features of the audio file, which results in the accuracy improvement. Additionally, normalization of the input features and audio segmentation largely improved the model performance.